Add passport and passport config, users controller, tied users and contacts

+++ service/contact.js
@@ -1,29 +1,43 @@
 const Contact = require("./schema/contact.js");
 
-const getContacts = async () => {
-  return await Contact.find();
-};
-
-const getContactById = async (contactId) => {
-  return await Contact.findById(contactId);
-};
-
-const removeContact = async (contactId) => {
-  return await Contact.findByIdAndRemove(contactId);
-};
-
-const addContact = async (body) => {
-  return await Contact.create(body);
-};
-
-const updateContact = async (contactId, body) => {
-  return await Contact.findOneAndReplace({ _id: contactId }, body, {
-    new: true,
+const getContacts = async (userId, query) => {
+  return await Contact.find({ owner: userId }).populate({
+    path: "owner",
+    select: "name email",
   });
 };
 
-const patchUpdateContact = async (contactId, body) => {
-  return await Contact.findByIdAndUpdate(contactId, body, { new: true });
+const getContactById = async (userId, contactId) => {
+  return await Contact.findById({ _id: contactId, owner: userId }).populate({
+    path: "owner",
+    select: "email",
+  });
+};
+
+const removeContact = async (userId, contactId) => {
+  return await Contact.findByIdAndRemove({ _id: contactId, owner: userId });
+};
+
+const addContact = async (userId, body) => {
+  return await Contact.create({ ...body, owner: userId });
+};
+
+const updateContact = async (userId, contactId, body) => {
+  return await Contact.findOneAndReplace(
+    { _id: contactId, owner: userId },
+    body,
+    {
+      new: true,
+    }
+  );
+};
+
+const patchUpdateContact = async (userId, contactId, body) => {
+  return await Contact.findByIdAndUpdate(
+    { _id: contactId, owner: userId },
+    body,
+    { new: true }
+  );
 };
 
 module.exports = {
